+++ changelog.md
@@ -113,6 +113,7 @@ Codion Change Log
 - DefaultFilterListModel.DefaultItemsBuilder.items() refactoring bug fixed, now allows null.
 - ListBuilder.ModelBuilder added.
 - SliderBuilder.ModelBuilder added.
+- ComboBoxBuilder.ModelBuilder added.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui